% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardElements.R
\name{createDashboardMatrix}
\alias{createDashboardMatrix}
\title{Function to extend dictionary for tableau matrices}
\usage{
createDashboardMatrix(dict, df)
}
\arguments{
\item{dict}{Extended dictionary from createExtendedDict}

\item{df}{interim dataframe to create tables from}
}
\value{
dashboard matrices
}
\description{
This function creates a df with all frequency tables, weighted if necessary
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{dictionary}
\keyword{microdata}
\keyword{sensata}
\keyword{tableau}
\keyword{tables}
