% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDf4GraphsCCV.R
\name{createDataGraphDumEduc}
\alias{createDataGraphDumEduc}
\title{Function to create data for edad and dummy graphs, final report CCV}
\usage{
createDataGraphDumEduc(df = CCVFinalData, originVar, levelToDummy = 1)
}
\arguments{
\item{df}{default: CCVFinalData}

\item{originVar}{Variable to be graphed}
}
\value{
Dataframe for graphs
}
\description{
This function creates graphData to create time series graph.
It also assumes the dataStructure of the CCVFinalData
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
