% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDf4GraphsCCV.R
\name{createDataGraphDumTs}
\alias{createDataGraphDumTs}
\title{Function to create data for time series for level selected for graphs, final report CCV}
\usage{
createDataGraphDumTs(df = CCVFinalData, ciudades, originVar, levelToDummy = 1)
}
\arguments{
\item{df}{default: CCVFinalData}

\item{originVar}{Variable to be graphed}
}
\value{
Dataframe for graphs
}
\description{
This function creates graphData to create time series graph.
The level required must be chosen with the position of the level, not the text.
It also assumes the dataStructure of the CCVFinalData
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
