% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDf4GraphsCCV.R
\name{createDataGraphNoReq}
\alias{createDataGraphNoReq}
\title{Function to create data for alt option as requirió for graphs, final report CCV}
\usage{
createDataGraphNoReq(df = CCVFinalData, originVar)
}
\arguments{
\item{df}{default: CCVFinalData}

\item{originVar}{Variable to be graphed}
}
\value{
Dataframe for graphs
}
\description{
This function creates graphData to create time series graph.
It assumes a 1-5 scale, values greater than 7 are missing. 6 are counted for the percentage.
It also assumes the dataStructure of the CCVFinalData
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
