% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDf4GraphsCCV.R
\name{createDataGraphPro2Vars}
\alias{createDataGraphPro2Vars}
\title{Function to create data for time series stacked for graphs, final report CCV}
\usage{
createDataGraphPro2Vars(df = CCVFinalData)
}
\arguments{
\item{df}{default: CCVFinalData}

\item{originVar1}{Variable to be graphed 1}

\item{originVar2}{Variable to be graphed 2}
}
\value{
Dataframe for graphs
}
\description{
This function creates graphData to create time series graph.
It assumes a 1-5 scale, values greater than 5 are missing.
It also assumes the dataStructure of the CCVFinalData
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
