% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExtendedDict.R
\name{createExtendedDict}
\alias{createExtendedDict}
\title{Function to extend dictionary for tableau matrices}
\usage{
createExtendedDict(
  dict,
  questions,
  disag,
  labels,
  topic,
  comments = NULL,
  abbrev = NULL
)
}
\arguments{
\item{dict}{data cleaned and ready for analysis}

\item{questions}{vector of variables (question identifiers) that should be kept from the dictionary}

\item{disag}{vector of variables (question identifiers) that will be on the cols of the tableau matrices, the disaggregations}

\item{labels}{labels for the disag variables}

\item{topic}{vector of topics for each question}

\item{comments}{(optional) vector of comment for each question, default NULL}

\item{abbrev}{(optional) vector of abbreviated text for questions, default NULL}
}
\value{
extended dictionary
}
\description{
This function creates a df with all frequency tables, weighted if necessary
}
\examples{
rawData <- sensataDataProg::sensataExample
Dict <- dictGenerator(
  df = rawData,
  questionPrefix = "",
  forceOrdered = "q_AB_NI_01")
questions <- c("q_AB_NI_01", "q_AB_EA_07", "q_AB_RA_01")
disag <- c("q_AB_CA_07", "q_AB_CA_09")
labels <- c("Marital status", "Student")
topic <- c("TEMA 1", "TEMA 2", "TEMA 2")
createExtendedDict(dict = Dict,
                   questions = questions,
                   disag = disag,
                   labels = labels,
                   topic = topic)
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{dictionary}
\keyword{microdata}
\keyword{sensata}
\keyword{tableau}
\keyword{tables}
