% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFreqTables.R
\name{createFreqTables}
\alias{createFreqTables}
\title{Function to create  crosstabulation tables}
\usage{
createFreqTables(
  df,
  rows,
  cols = NULL,
  weight = NULL,
  wide = FALSE,
  labels = "Total",
  percent = FALSE,
  addIdentifier = FALSE
)
}
\arguments{
\item{df}{data cleaned and ready for analysis}

\item{rows}{vector of variables that will be on the rows of the table}

\item{cols}{vector of variables that will be on the cols of the table}

\item{weight}{variable that identifies weights}

\item{wide}{logical, if TRUE expand for each value in cols}

\item{labels}{vector of labels for cols}

\item{percent}{logical, if TRUE table will have percentage instead of count}

\item{addIdentifier}{logical, if TRUE column Pregunta will include identifiers.}
}
\value{
Df of frequency tables
}
\description{
This function creates a df with all frequency tables, weighted if necessary
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{analysis}
\keyword{frequency}
\keyword{microdata}
\keyword{sensata}
\keyword{tables}
