% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraphData.R
\name{createGraphData}
\alias{createGraphData}
\title{Function to create data for graphs}
\usage{
createGraphData(
  df = intData,
  originVar,
  groupVar = "Edad",
  weigthVar = "ponde",
  totalColumn = T
)
}
\arguments{
\item{df}{default: intData}

\item{originVar}{Variable to be graphed. Default intData}

\item{groupVar}{Variable to be grouped by. Defaul Edad}

\item{totalColumn}{Should the total column be included in the data. Default: TRUE}

\item{nValues}{Number of values of originVar}

\item{weightVar}{Weight variable. Default ponde,}
}
\value{
Dataframe for graphs
}
\description{
This function creates weighted graphData to create most common graphs.
x axis is normallly the groupVar and y axis is the originVar.
if you don't want data to be weighted, use weightVar = NULL
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{graphs}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
