% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDf4GraphsAlimentacion.R,
%   R/createDf4GraphsAlimentacionV2.R
\name{createGraphData3OptionEdad}
\alias{createGraphData3OptionEdad}
\title{Function to create data for edad stacked for graphs, final report as}
\usage{
createGraphData3OptionEdad(df = asData, originVar, totalColumn = T)

createGraphData3OptionEdad(df = asData, originVar, totalColumn = T)
}
\arguments{
\item{df}{default: asData}

\item{originVar}{Variable to be graphed}

\item{totalColumn}{Should the total column be included in the data. Default: TRUE}
}
\value{
Dataframe for graphs

Dataframe for graphs
}
\description{
This function creates graphData to create stacked graph.
It assumes a 1-3 scale, values greater than 3 are missing.
It also assumes the dataStructure of the asFinalData

This function creates graphData to create stacked graph.
It assumes a 1-3 scale, values greater than 3 are missing.
It also assumes the dataStructure of the asFinalData
}
\examples{
TBD
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}

Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
