% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDf4GraphsAlimentacion.R,
%   R/createDf4GraphsAlimentacionV2.R, R/createDf4GraphsBPO-HGS.R,
%   R/createDf4GraphsBPO.R
\name{createGraphData4OptionGen}
\alias{createGraphData4OptionGen}
\title{Function to create data for gender stacked for graphs, final report as}
\usage{
createGraphData4OptionGen(df = asData, originVar, totalColumn = T)

createGraphData4OptionGen(df = asData, originVar, totalColumn = T)

createGraphData4OptionGen(df = asData, originVar, totalColumn = T)

createGraphData4OptionGen(df = asData, originVar, totalColumn = T)
}
\arguments{
\item{df}{default: asData}

\item{originVar}{Variable to be graphed}

\item{totalColumn}{Should the total column be included in the data. Default: TRUE}
}
\value{
Dataframe for graphs

Dataframe for graphs

Dataframe for graphs

Dataframe for graphs
}
\description{
This function creates graphData to create stacked graph.
It assumes a 1-4 scale, values greater than 4 are missing.
It also assumes the dataStructure of asData

This function creates graphData to create stacked graph.
It assumes a 1-4 scale, values greater than 4 are missing.
It also assumes the dataStructure of asData

This function creates graphData to create stacked graph.
It assumes a 1-7 scale, values greater than 7 are missing.
It also assumes the dataStructure of the asFinalData

This function creates graphData to create stacked graph.
It assumes a 1-7 scale, values greater than 7 are missing.
It also assumes the dataStructure of the asFinalData
}
\examples{
TBD
TBD
TBD
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}

Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}

Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}

Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
