% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGraphData.R, R/createGraphData5option.R
\name{createGraphData5option}
\alias{createGraphData5option}
\title{Function to create data for graphs}
\usage{
createGraphData5option(
  df = intData,
  originVar,
  groupVar = "Edad",
  totalColumn = T
)

createGraphData5option(
  df = intData,
  originVar,
  groupVar = "Edad",
  totalColumn = T
)
}
\arguments{
\item{df}{default: intData}

\item{originVar}{Variable to be graphed. Default intData}

\item{groupVar}{Variable to be grouped by. Defaul Edad}

\item{totalColumn}{Should the total column be included in the data. Default: TRUE}

\item{weightVar}{Weight variable. Default ponde}
}
\value{
Dataframe for graphs

Dataframe for graphs
}
\description{
This function creates weighted graphData to create most common graphs.
x axis is normallly the groupVar and y axis is the originVar.
if you don't want data to be weighted, use weightVar = NULL

This function creates graphData to create stacked graph.
It assumes a 1-5 scale, values greater than 5 are missing.
It also assumes the dataStructure of the asFinalData
}
\examples{
TBD
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}

Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{graphs}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
