% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDf4GraphsAlimentacion.R,
%   R/createDf4GraphsAlimentacionV2.R
\name{createGraphData5optionGen}
\alias{createGraphData5optionGen}
\title{Function to create data for gender stacked for graphs, final report as}
\usage{
createGraphData5optionGen(df = asData, originVar, totalColumn = T)

createGraphData5optionGen(df = asData, originVar, totalColumn = T)
}
\arguments{
\item{df}{default: asData}

\item{originVar}{Variable to be graphed}

\item{totalColumn}{Should the total column be included in the data. Default: TRUE}
}
\value{
Dataframe for graphs

Dataframe for graphs
}
\description{
This function creates graphData to create stacked graph.
It assumes a 1-5 scale, values greater than 5 are missing.
It also assumes the dataStructure of asData

This function creates graphData to create stacked graph.
It assumes a 1-5 scale, values greater than 5 are missing.
It also assumes the dataStructure of asData
}
\examples{
TBD
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}

Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
