% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDf4GraphsAlimentacion.R
\name{createGraphData8optionGen}
\alias{createGraphData8optionGen}
\title{Function to create data for gender stacked for graphs, final report as}
\usage{
createGraphData8optionGen(df = asData, originVar, totalColumn = T)
}
\arguments{
\item{df}{default: asData}

\item{originVar}{Variable to be graphed}

\item{totalColumn}{Should the total column be included in the data. Default: TRUE}
}
\value{
Dataframe for graphs
}
\description{
This function creates graphData to create stacked graph.
It assumes a 1-8 scale, values greater than 8 are missing.
It also assumes the dataStructure of asData
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
