% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDf4GraphsCuidado.R
\name{createGraphDataStackedGen}
\alias{createGraphDataStackedGen}
\title{Function to create data for gender stacked for graphs, final report CCV}
\usage{
createGraphDataStackedGen(df = cuidData, originVar)
}
\arguments{
\item{df}{default: cuidData}

\item{originVar}{Variable to be graphed}
}
\value{
Dataframe for graphs
}
\description{
This function creates graphData to create stacked graph.
It assumes a 1-5 scale, values greater than 5 are missing.
It also assumes the dataStructure of cuidData
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
