% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStackedList.R
\name{createStackedList}
\alias{createStackedList}
\title{Function to create data for graphs}
\usage{
createStackedList(varString, groupVar, weightVar = "ponde", totalColumn = F)
}
\arguments{
\item{varString}{Vector of variables to be graphed}

\item{groupVar}{Variable to be grouped by}

\item{weightVar}{Weight variable. Default ponde,}

\item{totalColumn}{Should the total column be included in the data. Default: TRUE}
}
\value{
Dataframe for graphs
}
\description{
This function creates weighted graphData and plots of the stacked variety (normally likerts).
if you don't want data to be weighted, use weightVar = NULL
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{data-cleaning}
\keyword{graphs}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
