% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFreqTables.R
\name{createTotalTables}
\alias{createTotalTables}
\title{Function to create Total tables}
\usage{
createTotalTables(df, rows, weight = NULL, percent = F, addIdentifier = F)
}
\arguments{
\item{df}{data cleaned and ready for analysis}

\item{rows}{vector of variables that will be on the rows of the table}

\item{weight}{variable that identifies weights}

\item{percent}{logical, if TRUE table will have percentage instead of count}
}
\value{
Df of frequency tables
}
\description{
This function creates a df with all frequency tables, weighted if necessary
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{analysis}
\keyword{frequency}
\keyword{microdata}
\keyword{sensata}
\keyword{tables}
