% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMnl.R
\name{predictMnl}
\alias{predictMnl}
\title{Function for predicting \% shares from a multinomial logit model
Based on chapter 13 of R for Marketing Research and Analytics https://link.springer.com/book/10.1007/978-3-319-14436-8
licensed with apache
Function for predicting \% shares from a multinomial logit model
Based on chapter 13 of R for Marketing Research and Analytics https://link.springer.com/book/10.1007/978-3-319-14436-8
licensed with apache}
\usage{
predictMnl(model, data)
}
\arguments{
\item{model}{mlogit object returned by mlogit()}

\item{data}{a data frame containing the set of designs for which you want to predict shares. Same format as the data used to estimate model.}
}
\value{
Dataframe with shares per level
}
\description{
Function for predicting \% shares from a multinomial logit model
Based on chapter 13 of R for Marketing Research and Analytics https://link.springer.com/book/10.1007/978-3-319-14436-8
licensed with apache
Function for predicting \% shares from a multinomial logit model
Based on chapter 13 of R for Marketing Research and Analytics https://link.springer.com/book/10.1007/978-3-319-14436-8
licensed with apache
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{analysis}
\keyword{conjoint}
\keyword{microdata}
\keyword{multinomial-logit}
\keyword{sensata}
