% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensataTheme.R
\name{sensata_palette}
\alias{sensata_palette}
\title{Palette for Sensata Theme}
\usage{
sensata_palette(
  palette = c("#003f5c", "#58508d", "#bc5090", "#ff6361", "#ffa600", "#C0C0C0")
)
}
\value{
A vector of colors
}
\description{
A custom color palette for Sensata theme.
}
\examples{
my_palette <- sensata_palette()
my_palette
}
