% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensataDataAnalysis-package.R
\docType{package}
\name{sensataDataAnalysis}
\alias{sensataDataAnalysis}
\title{sensataDataAnalysis: Analyzing sensata survey data}
\source{
\url{https://github.com/GaborioSensata/sensataDataAnalysis}
}
\description{
This package has the functions most commonly used to analyze survey data from Sensata platform

Este paquete tiene las funciones usadas comunmente para analizar los datos de encuestas de la plataforma de Sensata
}
\section{sensataDataAnalysis basic functions}{

LLENAR ESTA LISTA CON FUNCIONES PRINCIPALES
}

\section{Additional functions}{

LLENAR ESTA SECCIÓN CON OTROS PAQUETES
}

