% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sensataExAnalysis}
\alias{sensataExAnalysis}
\title{Sensata data example from Alimentación project}
\format{
A data frame with 2831 rows and 86 variables:
\describe{
\item{Group}{experimental group}
\item{DV}{experimenta dependent variable}
}
}
\source{
Alimentación saludable filtered by Colombia and only people with the experiment
}
\usage{
sensataExAnalysis
}
\description{
A dataset of survey 2831 surveys selected from Colombia Alimentación Saludable Data
sensataExAnalysis
}
\keyword{datasets}
