% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivityMnl.R
\name{sensitivityMnl}
\alias{sensitivityMnl}
\title{Function for creating data for a share - sensitivity chart
Based on chapter 13 of R for Marketing Research and Analytics https://link.springer.com/book/10.1007/978-3-319-14436-8
licensed with apache
Function for creating data for a share - sensitivity chart
Based on chapter 13 of R for Marketing Research and Analytics https://link.springer.com/book/10.1007/978-3-319-14436-8
licensed with apache}
\usage{
sensitivityMnl(model, attrib, base.data, competitor.data)
}
\arguments{
\item{model}{mlogit object returned by mlogit()}

\item{attrib}{list of vectors with attribute levels to be used in sensitivity}

\item{base.data}{: data frame containing baseline design of target product}

\item{competitor.data}{: data frame containing design of competitive set}
}
\value{
Dataframe with changes in share for sensitivity chart
}
\description{
Function for creating data for a share - sensitivity chart
Based on chapter 13 of R for Marketing Research and Analytics https://link.springer.com/book/10.1007/978-3-319-14436-8
licensed with apache
Function for creating data for a share - sensitivity chart
Based on chapter 13 of R for Marketing Research and Analytics https://link.springer.com/book/10.1007/978-3-319-14436-8
licensed with apache
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{analysis}
\keyword{conjoint}
\keyword{microdata}
\keyword{multinomial-logit}
\keyword{sensata}
