% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Summarizes data}
\usage{
summarySE(
  data = NULL,
  measureVar,
  weightsVar = NULL,
  groupVars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE,
  renameMean = FALSE
)
}
\arguments{
\item{data}{a data frame}

\item{measureVar}{the name of a column that contains the variable to be summarized}

\item{groupVars}{a vector containing names of columns that contain grouping variable(s)}

\item{na.rm}{a boolean that indicates whether to ignore NA's}

\item{renameMean}{a boolean that indicates if the mean column should be renamed}
}
\value{
Dataframe for graphs
}
\description{
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95\%)
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{analysis}
\keyword{metadata}
\keyword{microdata}
\keyword{sensata}
\keyword{summary-statistics}
