% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensataTheme.R
\name{theme_sensata}
\alias{theme_sensata}
\title{theme_sensata: Custom Theme for Sensata}
\usage{
theme_sensata()
}
\value{
A ggplot2 theme object
}
\description{
A custom theme for Sensata with custom color palette and fonts for ggplot2.
}
\examples{
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() +
  theme_sensata()
}
\seealso{
\code{\link[ggplot2]{theme}}

\code{\link[showtext]{showtext_auto}}

\code{\link[sysfonts]{font_add_google}}
}
