% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardElements.R
\name{createDashboardMatrix}
\alias{createDashboardMatrix}
\title{Function to create dashboard matrices}
\usage{
createDashboardMatrix(
  extDict,
  df,
  weight = NULL,
  addIdentifier = FALSE,
  totals = TRUE
)
}
\arguments{
\item{df}{interim dataframe to create tables from}

\item{weight}{variable that identifies weights, default NULL}

\item{addIdentifier}{logical, if TRUE column Pregunta will include identifiers.}

\item{totals}{logical, if TRUE includes total tables at the end of matrix}

\item{dict}{Extended dictionary from createExtendedDict}
}
\value{
dashboard matrices
}
\description{
This function creates a df with all frequency tables, weighted if necessary
}
\examples{
TBD
}
\author{
Gabriel N. Camargo-Toledo \email{gcamargo@sensata.io}
}
\keyword{dictionary}
\keyword{microdata}
\keyword{sensata}
\keyword{tableau}
\keyword{tables}
