% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensataTheme.R
\name{theme_sensata}
\alias{theme_sensata}
\title{Sensata ggplot theme with several color palettes}
\usage{
theme_sensata(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Sensata ggplot theme with several color palettes
}
\concept{themes fivethirtyeight}
